+++ SCHOOL_MANAGEMENT_SYSTEM/Models/WaterUsage.cs
@@ -16,6 +16,6 @@ namespace SCHOOL_MANAGEMENT_SYSTEM.Models
         public decimal prerecord { get; set; }
         public DateTime? currentdate { get; set; }
         public decimal currentrecord { get; set; }
-        public decimal price { get; set; }
+        public int price { get; set; }
     }
 }
